+++ src/basic_crawler.js
@@ -4,6 +4,7 @@ import log from 'apify-shared/log';
 import { isPromise, checkParamPrototypeOrThrow } from './utils';
 import AutoscaledPool from './autoscaled_pool';
 import RequestList from './request_list';
+import { RequestQueue } from './request_queue';
 
 const DEFAULT_OPTIONS = {
     maxRequestRetries: 3,
@@ -37,7 +38,8 @@ const DEFAULT_OPTIONS = {
  * ```
  *
  * @param {Object} options
- * @param {RequestList} options.requestList List of the requests to be processed.
+ * @param {RequestList} [options.requestList] List of the requests to be processed.
+ * @param {RequestList} [options.requestQueue] Queue of the requests to be processed.
  * @param {Function} options.handleRequestFunction Function that processes a request. It must return a promise.
  * @param {Function} [options.handleFailedRequestFunction=({ request, error }) => log.error('Request failed', _.pick(request, 'url', 'uniqueKey'))`]
  *                   Function to handle requests that failed more then option.maxRequestRetries times.
@@ -51,6 +53,7 @@ export default class BasicCrawler {
     constructor(opts) {
         const {
             requestList,
+            requestQueue,
             handleRequestFunction,
             handleFailedRequestFunction,
             maxRequestRetries,
@@ -65,20 +68,27 @@ export default class BasicCrawler {
         checkParamOrThrow(handleRequestFunction, 'opts.handleRequestFunction', 'Function');
         checkParamOrThrow(handleFailedRequestFunction, 'opts.handleFailedRequestFunction', 'Function');
         checkParamOrThrow(maxRequestRetries, 'opts.maxRequestRetries', 'Number');
-        // TODO: make this optional once we have the request queue ready.
-        checkParamPrototypeOrThrow(requestList, 'opts.requestList', RequestList, 'Apify.RequestList');
+        checkParamPrototypeOrThrow(requestList, 'opts.requestList', RequestList, 'Apify.RequestList', true);
+        checkParamPrototypeOrThrow(requestQueue, 'opts.requestQueue', RequestQueue, 'Apify.RequestQueue', true);
+
+        if (!requestList && !requestQueue) {
+            throw new Error('At least one of the parameters "opts.requestList" and "opts.requestQueue" must be provided!');
+        }
 
         this.requestList = requestList;
+        this.requestQueue = requestQueue;
         this.handleRequestFunction = handleRequestFunction;
         this.handleFailedRequestFunction = handleFailedRequestFunction;
         this.maxRequestRetries = maxRequestRetries;
 
         this.autoscaledPool = new AutoscaledPool({
-            workerFunction: () => this._workerFunction(),
             maxMemoryMbytes,
             maxConcurrency,
             minConcurrency,
             minFreeMemoryRatio,
+            handleTaskFunction: () => this._handleTaskFunction(),
+            isFinishedFunction: () => this._isFinishedFunction(),
+            isTaskReadyFunction: () => this._isTaskReadyFunction(),
         });
     }
 
@@ -92,36 +102,113 @@ export default class BasicCrawler {
     }
 
     /**
-     * Wrapper around handleRequestFunction that catches errors and retries requests.
+     * Fetches request from either RequestList or RequestQueue. If request comes from a RequestList
+     * and RequestQueue is present then enqueues it to the queue first.
      *
      * @ignore
      */
-    _workerFunction() {
-        const request = this.requestList.fetchNextRequest();
+    _fetchNextRequest() {
+        if (!this.requestList) return this.requestQueue.fetchNextRequest();
 
-        if (!request) return;
+        return this.requestList
+            .fetchNextRequest()
+            .then((request) => {
+                if (!this.requestQueue) return request;
+                if (!request) return this.requestQueue.fetchNextRequest();
 
-        const handlePromise = this.handleRequestFunction({ request });
-        if (!isPromise(handlePromise)) throw new Error('User provided handleRequestFunction must return a Promise.');
+                return this.requestQueue
+                    .addRequest(request, { forefront: true })
+                    .then(() => {
+                        return Promise
+                            .all([
+                                this.requestQueue.fetchNextRequest(),
+                                this.requestList.markRequestHandled(request),
+                            ])
+                            .then(results => results[0]);
+                    // If requestQueue.addRequest() fails here then we must reclaim it back to
+                    // the RequestList because probably it's not yet in the queue!
+                    }, (err) => {
+                        log.exception(err, 'RequestQueue.addRequest() failed, reclaiming request back to queue', { request });
 
-        return handlePromise
-            .then(() => {
-                this.requestList.markRequestHandled(request);
+                        // Return null so that we finish immediately.
+                        return this.requestList
+                            .reclaimRequest(request)
+                            .then(() => null);
+                    });
             })
-            .catch((error) => {
-                request.pushErrorMessage(error);
+    }
 
-                // Retry request.
-                if (request.retryCount < this.maxRequestRetries) {
-                    request.retryCount++;
-                    this.requestList.reclaimRequest(request);
-                    return;
-                }
+    /**
+     * Wrapper around handleRequestFunction that fetches requests from RequestList/RequestQueue
+     * then retries them in a case of an error etc.
+     *
+     * @ignore
+     */
+    _handleTaskFunction() {
+        const source = this.requestQueue || this.requestList;
 
-                // Mark as failed.
-                this.requestList.markRequestHandled(request);
+        return this._fetchNextRequest()
+            .then((request) => {
+                if (!request) return;
 
-                return this.handleFailedRequestFunction({ request, error });
+                const handlePromise = this.handleRequestFunction({ request });
+                if (!isPromise(handlePromise)) throw new Error('User provided handleRequestFunction must return a Promise.');
+
+                return handlePromise
+                    .then(() => source.markRequestHandled(request))
+                    .catch((error) => {
+                        request.pushErrorMessage(error);
+
+                        // Retry request.
+                        if (request.retryCount < this.maxRequestRetries) {
+                            request.retryCount++;
+
+                            return source.reclaimRequest(request);
+                        }
+
+                        // Mark as failed.
+                        return source
+                            .markRequestHandled(request)
+                            .then(() => this.handleFailedRequestFunction({ request, error }));
+                    });
+
             });
+    }
+
+    /**
+     * Returns true if some RequestList and RequestQueue have request ready for processing.
+     *
+     * @ignore
+     */
+    _isTaskReadyFunction() {
+        return Promise
+            .resolve()
+            .then(() => {
+                if (!this.requestList) return false;
+
+                return this.requestList.isEmpty();
+            })
+            .then((isRequestListEmpty) => {
+                if (!isRequestListEmpty || !this.requestQueue) return isRequestListEmpty;
+
+                return this.requestQueue.isEmpty();
+            })
+            .then(areBothEmpty => !areBothEmpty)
+    }
+
+    /**
+     * Returns true if both RequestList and RequestQueue have all requests finished.
+     *
+     * @ignore
+     */
+    _isFinishedFunction() {
+        const promises = [];
+
+        if (this.requestList) promises.push(this.requestList.isFinished());
+        if (this.requestQueue) promises.push(this.requestQueue.isFinished());
+
+        return Promise
+            .all(promises)
+            .then(results => _.all(results));
     }
 }